refactor(utils): remove dead code and dedupe helper functions
- Delete unused getLinksInText (no non-spec references) and its tests
- Extract canonical toggleInArray into utils/arrayHelpers.ts; have
  toggleArrayItem (eventMap) and toggleSelectedTag (forumListChannels)
  re-export it instead of redefining identical logic
- Dedupe truncateDescription: eventSeo now imports the canonical copy
  from discussionSeo; remove redundant test blocks
- Extract shared resolveRoleBadge into utils/roleBadges.ts; forum/server
  badge helpers now map its result onto their scoped string unions
