+++ utils/forumRoleBadges.ts
@@ -1,38 +1,14 @@
+import { resolveRoleBadge, type RoleBadgeParams } from '@/utils/roleBadges';
+
 export type ForumRoleBadge = 'forumAdmin' | 'forumMod' | null;
 
-type GetForumRoleBadgeParams = {
-  username?: string | null;
-  modProfileName?: string | null;
-  adminUsernames?: string[];
-  modUsernames?: string[];
-  modProfileNames?: string[];
-};
-
-const hasMatch = (value: string | null | undefined, candidates: string[]) => {
-  if (!value) {
-    return false;
-  }
-
-  return candidates.includes(value);
-};
-
-export const getForumRoleBadge = ({
-  username,
-  modProfileName,
-  adminUsernames = [],
-  modUsernames = [],
-  modProfileNames = [],
-}: GetForumRoleBadgeParams): ForumRoleBadge => {
-  if (hasMatch(username, adminUsernames)) {
+export const getForumRoleBadge = (params: RoleBadgeParams): ForumRoleBadge => {
+  const kind = resolveRoleBadge(params);
+  if (kind === 'admin') {
     return 'forumAdmin';
   }
-
-  if (
-    hasMatch(username, modUsernames) ||
-    hasMatch(modProfileName, modProfileNames)
-  ) {
+  if (kind === 'mod') {
     return 'forumMod';
   }
-
   return null;
 };
